+++ Applications/Pgan.PoracleWebNet.Api/Controllers/EggController.cs
@@ -7,6 +7,7 @@ using Pgan.PoracleWebNet.Core.Models;
 namespace Pgan.PoracleWebNet.Api.Controllers;
 
 [Route("api/eggs")]
+[RequireFeatureEnabled(DisableFeatureKeys.Raids)]
 public class EggController(IEggService eggService) : BaseApiController
 {
     private readonly IEggService _eggService = eggService;
@@ -30,7 +31,6 @@ public class EggController(IEggService eggService) : BaseApiController
         return this.Ok(egg);
     }
 
-    [RequireFeatureEnabled(DisableFeatureKeys.Raids)]
     [HttpPost]
     public async Task<IActionResult> Create([FromBody] EggCreate model)
     {
@@ -53,7 +53,6 @@ public class EggController(IEggService eggService) : BaseApiController
         }, result);
     }
 
-    [RequireFeatureEnabled(DisableFeatureKeys.Raids)]
     [HttpPut("{uid:int}")]
     public async Task<IActionResult> Update(int uid, [FromBody] EggUpdate model)
     {
@@ -100,7 +99,6 @@ public class EggController(IEggService eggService) : BaseApiController
         });
     }
 
-    [RequireFeatureEnabled(DisableFeatureKeys.Raids)]
     [HttpPut("distance/bulk")]
     public async Task<IActionResult> UpdateBulkDistance([FromBody] BulkDistanceRequest request)
     {
@@ -111,7 +109,6 @@ public class EggController(IEggService eggService) : BaseApiController
         });
     }
 
-    [RequireFeatureEnabled(DisableFeatureKeys.Raids)]
     [HttpPut("distance")]
     public async Task<IActionResult> UpdateAllDistance([FromBody] int distance)
     {